#13: another step of refactoring
