+++ src/gcd_engine.cpp
@@ -19,7 +19,7 @@
 */
 
 #include "gcd_engine.hpp"
-#include "m_motor_rpi.hpp"
+#include "driver_raspberry_pi.hpp"
 #include <list>
 #include <nlohmann/json.hpp>
 #include <regex>
