+++ src/i_Machine.hpp
@@ -18,8 +18,8 @@
 
 */
 
-#ifndef __MACHINE__HPP____
-#define __MACHINE__HPP____
+#ifndef __TP_MOTOR_I_MACHINE_HPP____
+#define __TP_MOTOR_I_MACHINE_HPP____
 
 #include "i_CoordTranslate.hpp"
 #include "i_MotorMoves_factory.hpp"
@@ -69,52 +69,6 @@ namespace motor {
         virtual void waitFinish() = 0;
     };
 
-    class Machine : public i_Machine {
-    protected:
-        std::shared_ptr<tp::coord::i_CoordTranslate> coordTranslator_;
-        std::shared_ptr<i_MotorMoves> motorMoves_;
-        Position minimalStepMm; // this is the positive vector for minimal step in direction (1,1,1,1) in steps
-
-        std::mutex m_csteps_;
-        Steps currentSteps_; // current position in steps
-
-        std::atomic_bool doBreak;
-
-    public:
-        void breakExecution();
-
-        // goes to selected position
-        void gotoXYZ(const Position& pos, double v, int selectedAxes = 0x0ff, double maxVelocityNoAccel = -1, double ddt = 100);
-
-        // enables or disables spindle
-        void spindleEnabled(bool _enabled);
-
-        // set steppers on or off
-        void steppersEnable(bool _enabled);
-
-        bool isSteppersEnabled();
-
-        // pause for given time
-        void pause(int ms_time);
-
-        // resets the position
-        void setPosition(const Position& pos);
-        // returns the current position
-        Position getPosition(bool sync = true);
-
-        // waits for commands to finish
-        void waitFinish();
-        // configuration of hardware layer
-        void setMotorMoves(std::shared_ptr<i_MotorMoves> _motorMoves);
-        // gets the coordinate system translation -- this will make smart pointers out of it
-        void setCoordinateSystem(std::shared_ptr<tp::coord::i_CoordTranslate> _coordTranslator);
-
-        int waitForEndstopTrigger();
-
-        std::array<unsigned char, 4> getEndstops();
-
-        Machine() { doBreak = false; }
-    };
 
 } // namespace motor
 } // namespace tp
